Test repeat script

+++ release/run_release_test.sh
@@ -44,19 +44,24 @@ esac
 shift
 done
 
+RAY_TEST_SCRIPT=${RAY_TEST_SCRIPT-ray_release/scripts/run_release_test.py}
 RAY_TEST_REPO=${RAY_TEST_REPO-https://github.com/ray-project/ray.git}
 RAY_TEST_BRANCH=${RAY_TEST_BRANCH-master}
 RELEASE_RESULTS_DIR=${RELEASE_RESULTS_DIR-/tmp/artifacts}
 
 export RAY_TEST_REPO RAY_TEST_BRANCH RELEASE_RESULTS_DIR
 
-pip uninstall -q -y ray
-pip install -q -r requirements.txt
-pip install -q -U boto3 botocore
-TMPDIR=$(mktemp -d -t release-XXXXXXXXXX)
-git clone -b "${RAY_TEST_BRANCH}" "${RAY_TEST_REPO}" "${TMPDIR}"
+if [ -z "${NO_INSTALL}" ]; then
+  pip uninstall -q -y ray
+  pip install -q -r requirements.txt
+  pip install -q -U boto3 botocore
+fi
 
-pushd "${TMPDIR}" || true
+if [ -z "${NO_CLONE}" ]; then
+  TMPDIR=$(mktemp -d -t release-XXXXXXXXXX)
+  git clone --depth 1 -b "${RAY_TEST_BRANCH}" "${RAY_TEST_REPO}" "${TMPDIR}"
+  pushd "${TMPDIR}" || true
+fi
 
 RETRY_NUM=0
 MAX_RETRIES=${MAX_RETRIES-3}
@@ -73,16 +78,25 @@ while [ "$RETRY_NUM" -lt "$MAX_RETRIES" ]; do
   if [ "$RETRY_NUM" -gt 1 ]; then
     # Sleep for random time between 30 and 90 minutes
     SLEEP_TIME=$((1800 + RANDOM % 5400))
+
+    if [ -n "${OVERRIDE_SLEEP_TIME}" ]; then
+      SLEEP_TIME=${OVERRIDE_SLEEP_TIME}
+    fi
+
     echo "----------------------------------------"
     echo "Retry count: ${RETRY_NUM}/${MAX_RETRIES}. Sleeping for ${SLEEP_TIME} seconds before retrying the run."
     echo "----------------------------------------"
-    sleep ${SLEEP_TIME}
+    sleep "${SLEEP_TIME}"
   fi
 
-  sudo rm -rf "${RELEASE_RESULTS_DIR}"/* || true
+  if [ -z "${NO_ARTIFACTS}" ]; then
+    sudo rm -rf "${RELEASE_RESULTS_DIR}"/* || true
+  fi
 
-  python ray_release/scripts/run_release_test.py "$@"
+  set +e
+  python "${RAY_TEST_SCRIPT}" "$@"
   EXIT_CODE=$?
+  set -e
   REASON=$(reason "${EXIT_CODE}")
   ALL_EXIT_CODES[${#ALL_EXIT_CODES[@]}]=$EXIT_CODE
 
@@ -91,7 +105,7 @@ while [ "$RETRY_NUM" -lt "$MAX_RETRIES" ]; do
     echo "Script finished successfully on try ${RETRY_NUM}/${MAX_RETRIES}"
     break
     ;;
-    7 | 9 | 10)
+    30 | 31 | 32 | 33)
     echo "Script failed on try ${RETRY_NUM}/${MAX_RETRIES} with exit code ${EXIT_CODE} (${REASON})."
     ;;
     *)
@@ -102,8 +116,10 @@ while [ "$RETRY_NUM" -lt "$MAX_RETRIES" ]; do
 
 done
 
-sudo rm -rf /tmp/ray_release_test_artifacts/* || true
-sudo cp -rf "${RELEASE_RESULTS_DIR}"/* /tmp/ray_release_test_artifacts/ || true
+if [ -z "${NO_ARTIFACTS}" ]; then
+  sudo rm -rf /tmp/ray_release_test_artifacts/* || true
+  sudo cp -rf "${RELEASE_RESULTS_DIR}"/* /tmp/ray_release_test_artifacts/ || true
+fi
 
 echo "----------------------------------------"
 echo "release test finished with final exit code ${EXIT_CODE} after ${RETRY_NUM}/${MAX_RETRIES} tries"
@@ -121,13 +137,17 @@ echo "----------------------------------------"
 REASON=$(reason "${EXIT_CODE}")
 echo "Final release test exit code is ${EXIT_CODE} (${REASON})"
 
-if [ "$EXIT_CODE" -gt 30 ] && [ "$EXIT_CODE" -le 40 ]; then
+if [ "$EXIT_CODE" -eq 0 ]; then
+  echo "RELEASE MANAGER: This test seems to have passed."
+elif [ "$EXIT_CODE" -gt 30 ] && [ "$EXIT_CODE" -le 40 ]; then
   echo "RELEASE MANAGER: This is likely an infra error that can be solved by RESTARTING this test."
 else
   echo "RELEASE MANAGER: This could be an error in the test. Please REVIEW THE LOGS and ping the test owner."
 fi
 
-popd || true
-rm -rf "${TMPDIR}" || true
+if [ -z "${NO_CLONE}" ]; then
+  popd || true
+  rm -rf "${TMPDIR}" || true
+fi
 
 exit $EXIT_CODE
